Moved context shortcut into a module
