+++ src/loggingex/context/__init__.py
@@ -13,6 +13,7 @@ from .exceptions import (
     ContextInvalidNameException,
 )
 from .filter import LoggingContextFilter
+from .shortcuts import context
 from .store import ContextStore
 
 
@@ -29,26 +30,3 @@ __all__ = (
     "LoggingContextFilter",
     "context",
 )
-
-
-class _ContextChangeShortcuts:
-    def __call__(self, *args, **kwargs) -> ContextChange:
-        """Create ContextChange object from positional and keyword arguments.
-
-        :param args: passed into remove.
-        :param kwargs: passed into update.
-        :return: new ContextChange object.
-        """
-        return ContextChange().remove(*args).update(**kwargs)
-
-    @staticmethod
-    def fresh(**kwargs) -> ContextChange:
-        """Create ContextChange object with fresh=True from keyword arguments.
-
-        :param kwargs: passed into update.
-        :return: new ContextChange object.
-        """
-        return ContextChange().fresh(True).update(**kwargs)
-
-
-context = _ContextChangeShortcuts()
